requirements.txt
openai>=1.30.0
pandas
python-dotenv>=1.0.1
requests>=2.32.0
tqdm>=4.66.0
